Move NodeRegistry impls to libsplinter
Moves the `NodeRegistry` implementations from splinterd to libsplinter,
since the existing implementations are not splinterd-specific.
